+++ packages/orq-rc/src/models/operations/deployments.ts
@@ -97,12 +97,12 @@ export const DeploymentsModelType = {
   Chat: "chat",
   Completion: "completion",
   Embedding: "embedding",
-  Vision: "vision",
   Image: "image",
   Tts: "tts",
   Stt: "stt",
   Rerank: "rerank",
   Moderations: "moderations",
+  Vision: "vision",
 } as const;
 /**
  * The modality of the model