+++ packages/orq-rc/src/models/operations/deploymentstream.ts
@@ -1428,7 +1428,6 @@ export const DeploymentStreamObject = {
   Chat: "chat",
   Completion: "completion",
   Image: "image",
-  Vision: "vision",
 } as const;
 /**
  * Indicates the type of model used to generate the response