+++ packages/orq-rc/src/models/operations/getoneprompt.ts
@@ -28,12 +28,12 @@ export const GetOnePromptModelType = {
   Chat: "chat",
   Completion: "completion",
   Embedding: "embedding",
-  Vision: "vision",
   Image: "image",
   Tts: "tts",
   Stt: "stt",
   Rerank: "rerank",
   Moderations: "moderations",
+  Vision: "vision",
 } as const;
 /**
  * The modality of the model