+++ packages/orq-rc/RELEASES.md
@@ -3599,3 +3599,13 @@ Based on:
 - [typescript v3.13.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.5 - packages/orq-rc
+
+## 2025-09-12 04:41:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.615.2 (2.698.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.6 - packages/orq-rc